Remove totals decoration (#67)
* Remove totals decoration

This only worked with a very specific setup. In the interest of simplicity I am removing it. I will create a ticket about possible hooks that could be in place to assign these. I might also just add some notes to the documentation about how this can be done with a custom serializer.

* Use default limit from config

* Documentation updates